Обновил список таблиц БД и добавил репозитории

+++ src/main/java/com/example/ncjavaproject/repositories/ValueRepository.java
@@ -3,5 +3,5 @@ package com.example.ncjavaproject.repositories;
 import com.example.ncjavaproject.models.Value;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ValueRepository extends CrudRepository<Value, Integer> {
+public interface ValueRepository extends CrudRepository<Value, Long> {
 }
